World Port Index
The World Port Index (Pub 150) contains the location and physical characteristics of, and the facilities and services offered by major ports and terminals world-wide. Entries are numbered geographically. In 2020, NGA undertook an effort to modernize the content and delivery of the World Port Index, resulting in additional and enhanced data fields, and a web application viewing platform. The modernization also implemented crowd-sourcing methods to validate and update the content.